Changed the return type of cycles3 and cycle4 functions from ulong to double.

+++ Random Networks Explorer/RegularHierarchicModel/RegularHierarchicNetworkAnalyzer.cs
@@ -60,7 +60,6 @@ namespace RegularHierarchicModel
 
         protected override double CalculateAverageClusteringCoefficient()
         {
-            // TODO 3 cycles in BigInteger
             double cycles3 = Count3Cycle(0, 0)[0], sum = 0, degree = 0;
             for (int i = 0; i < container.Size; ++i)
             {
@@ -71,16 +70,14 @@ namespace RegularHierarchicModel
             return 6 * cycles3 / sum;
         }
 
-        protected override long CalculateCycles3()
+        protected override double CalculateCycles3()
         {
-            // TODO get BigInteger result
-            return (long)Count3Cycle(0, 0)[0];
+            return Count3Cycle(0, 0)[0];
         }
 
-        protected override long CalculateCycles4()
+        protected override double CalculateCycles4()
         {
-            // TODO get BigInteger result
-            return (long)Count4Cycle(0, 0)[0];
+            return Count4Cycle(0, 0)[0];
         }
 
         protected override List<Double> CalculateEigenValues()
